\name{numSamples}
\alias{numSamples}
\title{Get the number of samples}
\description{
    This function extracts the number of samples from the \code{BEDMatrix}
    struct at the C level.
}
\usage{
numSamples(X)
}
\arguments{
    \item{X}{
        A \code{BEDMatrix} instance.
    }
}
\value{
    The number of samples.
}
