\name{numVariants}
\alias{numVariants}
\title{Get the number of variants}
\description{
    This function extracts the number of variants from the \code{BEDMatrix}
    struct at the C level.
}
\usage{
numVariants(X)
}
\arguments{
    \item{X}{
        A \code{BEDMatrix} instance.
    }
}
\value{
    The number of variants
}
